Adding Alt Text to images.
Added Alt text to images, it may not be needed feel free to ignore or
modify as you see fit.  At least having alt=“” would be a minimum since
you have the fig caption’s but the captions doesn’t talk about the
laptop / server / cellphone images in the picture.